2009-03-09	Joel Sherrill <joel.sherrill@OARcorp.com>
	PR 1376/filesystem
	* libcsupport/include/rtems/libio.h, libfs/src/dosfs/fat_file.c,
	libfs/src/dosfs/msdos.h, libfs/src/dosfs/msdos_dir.c,
	libfs/src/dosfs/msdos_file.c, libfs/src/imfs/deviceio.c,
	libfs/src/imfs/imfs.h, libfs/src/imfs/imfs_directory.c,
	libfs/src/imfs/memfile.c: lseek handlers should return off_t.

+++ cpukit/libcsupport/include/rtems/libio.h
@@ -82,7 +82,7 @@ typedef int (*rtems_filesystem_ioctl_t)(
   void          *buffer
 );
 
-typedef int (*rtems_filesystem_lseek_t)(
+typedef off_t (*rtems_filesystem_lseek_t)(
   rtems_libio_t *iop,
   off_t          length,
   int            whence
